+++ js/components/statistics-bar.js
