+++ js/screens/level-type-2-screen.js
@@ -6,7 +6,7 @@ import changeGameState from '../modules/change-game-state';
 
 import LevelTypeTwoView from '../views/level-type-2-view';
 import levelTypes from '../data/level-types';
-import greeting from "./greeting";
+import greeting from "./greeting-screen";
 
 export const levelTypeTwo = (gameState) => {
   const randomImages = generateRandomImages(2, ``);
