+++ js/screens/level-type-3-screen.js
@@ -4,22 +4,23 @@ import {checkAnswerTime} from '../modules/check-answer-time';
 import {renderLevel} from '../modules/render-level';
 import changeGameState from '../modules/change-game-state';
 
-import LevelTypeOneView from '../views/level-type-1-view';
+import LevelTypeThreeView from '../views/level-type-3-view';
 import levelTypes from '../data/level-types';
-import greeting from "./greeting";
+import greeting from "./greeting-screen";
 
-export const levelTypeOne = (gameState) => {
-  const randomImages = generateRandomImages(1, ``);
-  const screen = new LevelTypeOneView(gameState, randomImages);
+export const levelTypeThree = (gameState) => {
+  const randomImages = generateRandomImages(3, `paint`);
+  const screen = new LevelTypeThreeView(gameState, randomImages);
+
 
   screen.onBackButtonClick = () => {
     showScreen(greeting);
   };
 
-  screen.onRadioChangeHandler = (isAnswerCorrect) => {
+  screen.onButtonClickHandler = (isCorrectAnswer) => {
     const generatedLevelType = getRandomElement(levelTypes).type;
 
-    changeGameState(isAnswerCorrect, checkAnswerTime(50)); // temporarily static value
+    changeGameState(isCorrectAnswer, checkAnswerTime(20));
     renderLevel(generatedLevelType, gameState);
   };
