+++ js/screens/rules-screen.js
@@ -5,6 +5,8 @@ import RulesView from '../views/rules-view';
 
 export default class RulesScreen {
   constructor() {
+    this.screenContent = new RulesView();
+    this.header = new HeaderView();
     this.content = this.renderContent();
   }
 
@@ -13,22 +15,20 @@ export default class RulesScreen {
   }
 
   renderContent() {
-    const header = new HeaderView();
-    const screenContent = new RulesView();
     const container = document.createElement(`div`);
 
-    container.appendChild(header.element);
-    container.appendChild(screenContent.element);
+    container.appendChild(this.header.element);
+    container.appendChild(this.screenContent.element);
 
     return container;
   }
 
   init() {
-    this.content.onBackButtonClick = () => {
+    this.header.onBackButtonClick = () => {
       Application.showScreen(ScreenName.GREETING);
     };
 
-    this.content.onFormSubmit = (player) => {
+    this.screenContent.onFormSubmit = (player) => {
       Application.showGame(player);
     };
   }
