+++ js/views/statistics-view.js
@@ -1,6 +1,6 @@
 import AbstractView from './abstract-view';
 import renderHeader from "./header-view";
-import statisticsBar from '../screens/statistics-bar';
+import statisticsBar from '../components/statistics-bar';
 import footerTemplate from "../components/footer";
 
 import {GamePoint} from '../data/constants';
